pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto_hft_tool"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "A crypto HFT tool for statistical arbitrage"
requires-python = ">=3.9"
dependencies = [
    "ccxt",
    "websockets",
    "pandas",
    "numpy",
    "python-dotenv",
    "pytest",
] 